Make sure that we don't try to trace all of memory by mistake.
JamVM likes to trace from the current stack frame up to the known top of
the stack to find references to Java objects on the stack.
Unfortunately, our stacks are not always contiguous when we're handling
a ccall.  This is, in particular, the case when invoking the security
manager.

For now, simply don't mark the thread as blocking, so the GC thread has
to wait for longer.